+++ packages/components/__test_utils__/withProviders/types.ts
@@ -1,0 +1,6 @@
+import type { RelayMockEnvironment } from 'relay-test-utils/lib/RelayModernMockEnvironment'
+
+export type WithProvidersOptions = {
+  environment: RelayMockEnvironment
+  [key: string]: any
+}