+++ packages/components/modules/comments/CommentItem/CommentsReplies/index.tsx
@@ -79,7 +79,8 @@ const CommentsReplies: FC<CommentsRepliesProps> = ({
       <Virtuoso
         useWindowScroll
         data={comments}
-        overscan={NUMBER_OF_COMMENTS_TO_LOAD_NEXT}
+        // TODO: using overscan can cause Maximum call stack size exceeded error
+        // overscan={NUMBER_OF_COMMENTS_TO_LOAD_NEXT}
         itemContent={(_index, comment) => renderCommentItem(comment)}
         components={{
           Header: renderHeader,