+++ packages/components/modules/comments/CommentsList/index.tsx
@@ -69,7 +69,8 @@ const CommentsList: FC<CommentsListProps> = ({
         <Virtuoso
           useWindowScroll
           data={comments}
-          overscan={NUMBER_OF_COMMENTS_TO_LOAD_NEXT}
+          // TODO: using overscan can cause Maximum call stack size exceeded error
+          // overscan={NUMBER_OF_COMMENTS_TO_LOAD_NEXT}
           itemContent={(_index, comment) => renderCommentItem(comment)}
           components={{
             Header: renderHeader,