+++ packages/components/modules/messages/index.ts
@@ -1,7 +1,18 @@
+export { default as MessageItem } from './MessageItem'
+export type * from './MessageItem/types'
+
+export { default as MessagesList } from './MessagesList'
+export type * from './MessagesList/types'
+
+export { default as MessageRoom } from './MessageRoom'
+export type * from './MessageRoom/types'
+
 export { default as SendMessage } from './SendMessage'
 export type * from './SendMessage/types'
 
 export * from './graphql/mutations/SendMessage'
-
 export * from './graphql/queries/MessageItem'
 export * from './graphql/queries/MessagesList'
+export * from './graphql/queries/MessageRoomQuery'
+
+export { default as useMessagesListSubscription } from './graphql/subscriptions/useMessagesListSubscription'