+++ packages/components/modules/profiles/ProfilePopover/AddProfileMenuItem/types.ts
@@ -1,0 +1,3 @@
+export interface AddProfileMenuItemProps {
+  addNewProfileLabel?: string
+}