+++ packages/components/modules/profiles/context/CurrentProfileProvider/constants.ts
@@ -1,0 +1,7 @@
+import type { CurrentProfileState } from './types'
+
+export const CURRENT_PROFILE_STORAGE_KEY = 'currentProfile'
+export const INITIAL_CURRENT_PROFILE_STATE: CurrentProfileState = {
+  profile: undefined,
+  userId: undefined,
+}