+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @baseapp-frontend/design-system
 
+## 0.0.17
+
+### Patch Changes
+
+- Storybook cleanup
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/utils@3.0.3
+
+## 0.0.15
+
+### Patch Changes
+
+- Add the AddIcon to the icons library
+
 ## 0.0.14
 
 ### Patch Changes